feat(source): move transformer to configurator
